config: espressobin-88f3720: move env to the end of the 4MB boot device
The size of flash_image.bin with the tee binary exceeds 1.5MB.
The tee binary size can grow up to 2MB, and without the environment
relocation leads to a boot image corruption upon running saveenv.
Relocate the u-boot environment to the end of a minimum-sized
boot device (flash or eMMC boot partition of 4MB).

Signed-off-by: Zhoujie Wu <zjwu@marvell.com>
Signed-off-by: wangwen <wangwen@marvell.com>
Signed-off-by: Konstantin Porotchkin <kostap@marvell.com>
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
Tested-by: Igal Liberman <igall@marvell.com>