dm6291: add backup boot support
These are presumably the functions which were added to U-Boot to check
integrity of main kernel image and boot a backup kernel in case of a
failure. Essentially this is a diff between uboot/u-boot-dm6291-new and
uboot/u-boot-dm6291-new-nor at https://github.com/tonystark3188/6291.git,
with small changes to make it behave like factory U-Boot for ZoomGo ZX10.

Signed-off-by: Tomasz Maciej Nowak <tmn505@gmail.com>